READ ME FILE FOR REPOSITORY
# COLLLABORATION
This repository will contain the files of my second project of DT Batch(JAVA ENTERPRISE WITH DEVOPS DEVELOPER). 
All the files in the repository will be compressed to .rar format. 
So that you can download it and extract it to your workspace of Eclipse or Netbean and run the project.
All the update have dats in thier name. Further upload will also have this name format.